+++ changelog.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.2.3] - 2018-10-29
+
+### Added
+
+- Added a new method `setScriptInfo` to set script info after constructor.
+
+### Fixed
+
+- Relative paths in constructor options (e.g. `collection`) should resolve relative to option
+  `resolveFrom`.
+
 ## [0.2.2] - 2018-10-28
 
 ### Added
@@ -153,7 +164,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Initial release
 
-[Unreleased]: https://github.com/revam/node-script-loader/compare/v0.2.2...HEAD
+[Unreleased]: https://github.com/revam/node-script-loader/compare/v0.2.3...HEAD
+[0.2.3]: https://github.com/revam/node-script-loader/compare/v0.2.2...v0.2.3
 [0.2.2]: https://github.com/revam/node-script-loader/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/revam/node-script-loader/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/revam/node-script-loader/compare/v0.1.2...v0.2.0